requirements.txt
--index-url https://pypi.python.org/simple/

git+https://github.com/MapStory/composer.git@master#composer
git+https://github.com/boundlessgeo/django-exchange-maploom.git@master#django-exchange-maploom
git+https://github.com/boundlessgeo/geonode@exchange/1.2.0#egg=geonode
git+https://github.com/GeoNode/django-osgeo-importer@master#egg=django-osgeo-importer
git+https://github.com/boundlessgeo/exchange-documentation@master#egg=django-exchange-docs
git+https://github.com/venicegeo/nearsight@master#egg=nearsight
numpy==1.11.3
django-geonode-client==0.0.23
dj-database-url==0.4.2
celery[redis]==3.1.18
django-storages==1.1.8
boto==2.38.0
waitress==0.9.0
whitenoise==3.2
django-cors-headers==1.1.0
django-classification-banner==0.1.5
django-solo==1.1.2
django-colorfield==0.1.10
psycopg2==2.6.1
python-ldap==2.4.25
django-auth-ldap==1.2.8
GDAL==2.1.0
supervisor==3.3.1
python-resize-image==1.1.10
django-flat-theme==1.1.3
flake8==3.0.4
elasticsearch-dsl==5.0.0
social-auth-app-django==1.1.0
geopy==1.11.0
cx_Oracle==5.2.1
mock==2.0.0